profile service renamed to profile repository
